requirements.txt
opencv-python==4.10.0.84
ultralytics==8.3.30
python-telegram-bot==21.6
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
urllib3<2
